requirements.txt
astroid==2.4.2
cachetools==4.2.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-texttospeech==2.2.0
google-cloud-translate==3.0.2
googleapis-common-protos==1.52.0
grpcio==1.34.0
httplib2==0.18.1
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
libcst==0.3.16
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
Pillow==8.0.1
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
